MODE-1065 Corrected RepositoryService's observer to narrow the focus
The RepositoryService component (owned by the ModeShapeEngine and the JcrEngine subclass) has a change observer
that it uses to monitor the configuration repository for changes. Unfortunately, the observer was looking
for all changes under '/' and attempted to re-read the content to create a RepositorySource for each.
Normally this works well and handles the case where new sources are added or existing ones are modified or
removed. However, as node types defined via the configuration file are registered (by the JcrEngine),
these new node types are stored within the configuration repository. This results in changes that are picked
up by the RepositoryService's observer, and for each change (e.g., each new node type definition,
property definition, child node definition, etc.) the observer is re-reading the configuration content.
This is causing a significant number of unnecessary reads.

The problem is that the RepositoryService's observer is not checking that the path of the changed nodes
are below the parent of the "source" nodes. Once this is fixed, the observer ignores all changes except
those that are truly applicable to the "source" nodes, and this shortens the start-up time of the
JcrEngine (when there are a lot of custom node types) by roughly one-half!