MODE-1096, MODE-880 Support building with Maven 3, migrate from JBoss IDTrust to PicketBox
Changed the parent POM and quite a few module POM files to support building with Maven 3. Maven 2 is no longer supported.
As part of the change, several artifacts required by JBoss IDTrust could not be resolved in the JBoss Maven repository,
so IDTrust was replaced with PicketBox (which is the re-branded and updated IDTrust project). This required changes
to quite a few POMs and test cases, as the classes for the JAAS immplementation changed names and packages.

At this point, all unit and integration tests pass, and all assemblies appear to be generated correctly.
Incidentally, build times appear to be approximately 1/3 faster.