MODE-1102 Added support for the JPA connector to not validate the database schema
Changed the JpaSource class to allow a new "disabled" value for the "autoGenerateSchema" property that will not
set the Hibernate "hibernate.hbm2ddl.auto" configuration setting. This is now the preferred value for a production
system, since it means Hibernate will do no validation of the database schema. Best practice for production systems
is to have the DB admin explicitly manage the database schema, and this new mode allows the JPA connector to simply
assume that the database matches.

The Reference Guide's chapter describing the JPA connector was updated to reflect the new "autoGenerateSchema" property.

All unit and integration tests pass.