MODE-1104 Changed build to work in a single pass
Prior to this change, someone running a ModeShape build for the first time had to run two separate commands:
'mvn -Pintegration install' and 'mvn -Passembly install'. It also used aggregator goals for JavaDoc and
assemblies, which fork a new build process for each invocation and which can result in a build that does
double work (if they're included in the main build process). Plus, because the JavaDoc and assembly usages
were in the parent POM, the JavaDoc and assemblies were produced before the modules were compiled; on a machine
in which 'mvn install' was not yet run, this caused an error.

Now, the assembly operation is defined in a child module called 'modeshape-distribution' and the assembly
descriptors are referenced as being on the classpath since they're now defined in a new
'modeshape-assembly-descriptors' module. The end result is that, on a fresh machine on which ModeShape
has not yet been built, the source for ModeShape can be downloaded and completely built using a single
command: 'mvn -Passembly install'.

This commit moves quite a few files around and restructures a number of the POM files, but does not affect
any of the Java source code.