MODE-1106 Corrected/improved the examples' POMs
The parent POM for the examples was not correct, and was using different versions of some of the dependencies.
It was also duplicating a lot of the dependencies and other information inherited from the new 'modeshape-parent'
POM (which it was not doing before). Also, the two assembly descriptors were corrected to include all of the
dependencies' JARs without the Maven warning. And the examples' parent POM no longer produces any JARs or output.