MODE-1105 Corrected logic to import graph from classpath resource
The logic of importing Graph content from an XML file on the classpath was not correct,
and was always closing the stream before anything could be read from the stream.
Added a test case to verify the reported behavior, and then verified that moving
the 'stream.close()' call to the correct location works as expected.

All unit and integration tests pass with these changes.