MODE-1109 Corrected the RemoteInfinispanSource tests to work on IPv6.
When creating the HotRod server and client, the test code now dynamically looks up the address of the
local host and uses that for configuration. This works whether the local host has an IPv4 or IPv6
address. Also, some of the tests were not properly shutting down the client before the server,
which would occasionally fail if the server happened to shut down before the client.

All unit and integration tests pass with these changes.