MODE-1110 Corrected SelectComponent class to prevent NullPointerException.
The problem is in the SelectComponent.satisfiesConstraints(...) method and how it evaluates each
of the constraints upon the tuples returned from the source. In particular, the method assumed
an analyzer was present with the ISDESCENDANTNODE(selector,path) constraint.

The fix was to use the analyzer if not null, or if null to use the Location value in the tuples.

All unit and integration tests pass with these changes.