MODE-1111 Added singleton JNDI naming context and factory implementation, useful for testing
Added a new JNDI naming context implementation and an InitialContextFactory implementation that may be of use
during testing of services that use JNDI to register objects by name. These new classes result in a singleton
naming context, meaning that even when services are configured and register objects into a naming context,
the test case can access the same naming context instance to find the registered objects.