MODE-1099 Added ability for clients to publish files as being versioned
Previously, the REST client API would publish as file by creating an "nt:file" node
with the standard properties and "jcr:content" child node, but would never create nodes that
were versionable. This change adds a method to the IRestClient interface that allows a client
to specify whether the published file should be versioned, in which case the "mix:versionable"
mixin is added to the "nt:file" node's list of "jcr:mixinTypes".