MODE-1111 Merged the MockInitialContext previously in the test code with the new SingletonInitialContext
The new SingletonInitialContext and SingletonInitialContextFactory classes now in the 'modeshape-common'
module now supersede the MockInitialContext and MockInitialContextFactory classes that were previously
in the test code for that module. This means it is now possible to reuse these classes in code that
depends on ModeShape (without requiring a dependency on the test JAR).