MODE-1112 A temporary ZIP file used in the JBoss AS kit is no longer installed to Maven
The build process created a "jbossas-{version}-{version}-dist.zip" file that was used as an
intermediate file during assembly of the real "modeshape-jbossas-{version}-dist.zip" file.
Not only was the filename incorrect (the version was in the name twice), but the file should
not have been installed into the local repository. These changes correct the name
and prevent the installation.