MODE-1113 Corrected Repository initialization so that existing node types are read properly
The JcrRepository initialization logic was corrected to read any node types defined in prior executions
of the ModeShape engine are read from the "/jcr:system/jcr:nodeTypes" content. When the engine
starts up and initializes the RepositoryNodeTypeManager, it first reads any existing node types
in the system content (this is the new behavior), then registers any node type definitions that are
referenced in the configuration file, and then registers all of the built-in node type definitions.
In this way, the built-in definitions can never be overridden. This behavior also means that
if the configuration file references some CND files with user-defined node type definitions,
these definitions will override any changes that were made to those definitions before the last
shutdown.

These changes work for clustered configurations. All unit tests and integration tests pass,
including several new tests that verify this new behavior.