MODE-1114 Corrected import (and replace) behavior when imported content includes versionable nodes
Added several new integration tests to try and replicate the problem. One of these tests did replicate it:
the trick was to import an XML file (with referenceable nodes and versionable nodes), to re-import the
same XML file while replacing content with duplicate UUIDs, and then checking out, changing, and checking
in one of the versionable nodes (that is a descendant of another versionable node that was imported).
The checkin causes the problem because it attempted to resolve the 'jcr:predecessors' REFERENCE property,
which contains a UUID that does not correspond to a node in the repository.

After quite a bit of debugging (almost two days), I was able to figure out what was actually going on.
Basically, when the second import is happening and is importing a node with the same UUID as an existing
node, the existing node must be replaced (e.g., removed). If that existing node is also versionable,
then there is a version history and this was not being removed. Removing these version histories did not
work and resulting in what appeared to be invalid, since it prevents recovering of previously-versioning
information.

A second approach was more successful: leave the version history and remove the "mix:versionable" properties
from the newly-imported nodes (as was being done already). The JcrVersionManager code that initializes
the version history was always generating a new UUID for the base version, even though this new UUID is
not written to the version history if the version history already exists. (The JcrVersionManager code
creates a node with the UUID only if the node is absent.) This is the fundamental source of the problem.
Simply re-reading the root version's UUID and using it for the "jcr:baseVersion" property and in the
"jcr:predecessors" property appears to work.

The new integration tests test a number of permutations of operations on both in-memory and JDBC
connectors, and these tests pass - as do all other unit and integration tests. Therefore, this appears
to fix the issue, and results in identical behavior and version history content whether or not the
content is imported multiple times. In other words, the behavior when importing multiple times and modifying
the versioned content (using checkout, modify, checkin) is the same as importing once and doing the
same modifications.

However, during verification of the results, MODE-1117 was discovered. Although not technically related
to this issue, it does cause the "jcr:predecessors" and "jcr:successors" properties to be incorrect.