MODE-1117 Added sandbox project for testing JCR reference implementation behavior
Created a new 'modeshape-test-reference-impl' project under the 'sandbox' directory for
testing the JCR reference implementation behavior. The initial test verified that the
proposed change in behavior for MODE-1117 is indeed similar to the behavior of the reference
implementation.

Note that this new project is not automatically included in the standard builds, but
can be invoked via the 'sandbox' assembly. Also, this project does not build or install
any JARs.