MODE-1117 Corrected the logic that sets the 'jcr:predecessors' and 'jcr:successors' properties
Corrected the JcrVersionManager logic upon checkout to set the 'jcr:predecessors' property to the
'jcr:baseVersion' property value. Prior to this fix, checkout was also adding all of the values
from the base version's 'jcr:predecessors'; the net result of that was that the 'jcr:predecessors'
property always had all prior versions. See MODE-1117 for a more detailed explanation of the
problem.

Once JcrVersionManager was corrected (as described above), two TCK tests failed because of the
logic in the JcrVersionHistoryNode.removeVersion(...) method. Corrected this logic to more correctly set
the 'jcr:predecessors' on each of the removed version's successor nodes, and to correctly set
the 'jcr:successors' on each of the removed version's predecessor nodes.

After these changes, all unit and integration tests (including those added in MODE-1114) pass.