MODE-1119 Improved handling of errors in configuration
Changed the way that errors in the repository configurations are handled. Problems with
the initial content file location are reported with better messages (including that the
file cannot be found or might be empty). Also added a second "start" method on the
JcrEngine class that will automatically start all repositories in the configuration,
and will record all problems during repository initialization in the engine's problems.
This is necessary because the repository configurations are not really validated or used
until the repository is accessed; this new method provides a way to prestart all repositories
and, in the process, validate their configurations.