MODE-1112 Corrected distribution POM
The modeshape-distribution POM file was causing an error during the assembly of the module's JAR.
Unfortunately, I can turn off the creation of the JAR, but somehow the assembly is trying to
use the generated classes and resulting in an error. I've found no way of shutting this off.

Therefore, this fix changes the POM to create the 'modeshape-distribution-<version>.jar' file
and install it into the local repository. Note that this JAR file only contains the manifest.