MODE-1130 Corrected how the REST service updates versionable files
Added the test provided by Dan, and added a second test that was very similar (but not versionable). The versionable test failed, as Dan reported.

The REST service was not properly using the JCR 'checkout, modify, save, checkin' pattern. Instead, it was using 'checkout, modify, checkin, save'. Obviously this was wrong.

Once this was fixed (and a few changes were made to do only one session.save even if multiple versionable nodes are updated in one request), the new tests all work fine and all unit and integration tests pass.