MODE-1135 Added function framework to the Graph API
Added a new function framework to allow Graph users to write custom logic that will be executed within the connector when processing requests.

To use this new framework, a Graph users creates a concrete subclass of the Function abstract class, which has just one 'run(FunctionContext)' method, and submits it to the connector using Graph.applyFunction(...) or Graph.Batch.applyFunction(...). Because the Function.run(...) method will be called by the connector, the implementation can create and execute other requests (such as reading, creating, updating, or removing nodes and/or properties) intermixed with custom logic.

When the Graph.applyFunction(...) or Graph.Batch.applyFunction(...) methods are called, the method encapsulates the function, its inputs, the graph location under which the function will be applied, and the workspace name into an instance of FunctionRequest, which is a concrete subclass of Request. This FunctionRequest essentially encapsulates everything necessary to invoke the function at a later point. The FunctionRequest is even serializable.

Several new unit tests were added to the 'modeshape-graph' module, verifying that the new framework works as expected. Additionally, all other unit and integration test pass.