MODE-1134 Used graph functions to implement version history initialization
The new graph function feature (MODE-1135) was used to re-implement the logic of initializing the version history of a node. This logic required updating (with a lot of 'if absent' options) and re-reading the results. Using functions, all this logic of initializing (or reusing in the case of a copy or import) the version history under '/jcr:system/jcr:versionStorage' can be done within the connector's processor. And doing that means it's possible to have smarter logic that updates the content appropriately based upon what's already there.
The JcrVersionManager was also changed throw out of the cache only the subset of information that was changed. In some cases, this is still the entire version storage, but in other cases it is a much smaller subgraph. Additionally, the code is slightly less reliant upon the cached information, meaning the information is loaded less frequently to begin with.
Changed the JcrContentHandler class (using in importing) to use the new function to initialize (or reuse existing) version storage (in batch mode) for the 'mix:versionable' nodes created by the importer. Prior to this, the version storage was updated for each versionable node in separate calls, so this change should scale significantly better as there will be at most one system batch and one workspace batch. (The system and workspace changes cannot be easily combined, as they are to different workspaces; this may be done as part of other performance improvement steps.)
The JcrVersionManager's checkin method was also changed to utilize a function to more efficiently checkin the node. Note that per the spec, checkins are not recursive.
All unit and integration tests pass with these changes.