MODE-1137 Corrected handling of unicode characters in export/import
There were a couple of minor problems, and I think all should be cleared up.

First, the StreamingContentHandler should allow using different encodings, so this class was cleaned up to allow setting the encoding in the constructor and using the correct encoding when writing to the stream. The class is now using an OutputStreamWriter (initialized with the correct encoding) rather than writing the bytes encoded for each call. This should be a bit faster. Also, the reported issue won't always surface, because prior to this fix the code depended upon the platform's default encoding. Only those users with a default encodings other than "UTF-8" would see the problem. The fix should alleviate this issue.

Second, while testing it was discovered that we are always encoding the space character. Contrary to Section 6.4.4 of the JSR-170 specification, the reference implementation never encodes spaces in values. Therefore, this change removes the space character from the list of characters that are encoded upon export.

Third, several new unit tests that use Unicode content were added to the set of export/import round-trip cases, verifying the correct behavior.