MODE-1137 Additional fixes to the import/export behavior
Although none of the unit or integration tests were failing during the Maven build, several import/export related tests were failing within Eclipse. It turns out that some of the TCK tests are expecting spaces to be escaped in the document view, so this test reverts the prior commit's change on this behavior.