MODE-1140 Corrected builds to again run the JUnit 3 unit tests
When we changed to Maven 3, we also upgraded a number of the Maven plugins, in part because Maven 3 encourages specifying each of the version numbers for the plugins to keep your builds stable. When we switched from Surefire 2.6 to 2.7.1, I failed to notice that Surefire 2.7 determines the test classes to run in a vastly different way than 2.6, and one that will purposefully exclude all JUnit 3 test classes. Unfortunately for us, the JCR TCK unit test suites are all JUnit 3, and this means we haven't been running the JCR TCK unit tests (in any configuration) since the move to Maven 3 (and commit f3a8a09a3bb on Feb 22)!

Fixing this was painful and not obvious. It basically boils down to explicitly telling Surefire to use an older JUnit 4 provider (rather than a new "junit-47" provider that does the funky determination).

With these changes, all unit and integration tests pass, including all of the TCK tests. The time to run a full build is a bit faster than with Maven 2, but it's not quite the massive speedup that we thought we got. Oh well.

It is also possible to downgrade Surefire to 2.6, but I thought it better to keep Surefire on the latest possible version (which has performance and size improvements over 2.6).