MODE-1141 Corrected VersionManager.checkin(...) logic to not duplicate versionable nodes in history
As noted in the JIRA issue, the JcrVersionManager.checkin(...) method calls the versionNodeAt(...) recursive method, which copies the versionable state into the new version in version history. This method should *not* be walking children when the child is 'mix:versionable' and has an OPV on the child node definition of VERSION.

This change simply implements this behavior by returning from the method rather than breaking (the children are walked after the break).

All unit and integration tests pass with this minor change.