MODE-1143 Upgrade to Lucene 3.1
Changed the versions in the 'modeshape-search-lucene' module from 3.0.2 to 3.1.0, although the regex library is still at 3.0.3. The snowball library was incorporated into the contrib library, so it was removed.

Unfortunately, there are a lot of deprecations in the Lucene API, where some of the deprecated interfaces are still required. It looks like these are merely a step forward, and future releases will make additional deprecations/changes.

All unit and integration tests pass.