MODE-1144 Corrected logic to convert XPath queries to JCR-SQL2
First added a unit test in JcrQueryManagerTest to execute the same XPath query, and replicated the error. After a bit of searching, the problem turned out to be the generated JCR-SQL2 query was incorrect (used a different alias in the second criteria). Thus the problem was likely in the XPathToQueryTranslator, so another unit test was added to XPathToQueryTranslatorTest, which also failed.

After a bit of investigation, discovered that the processing of the criteria was incorrectly switching table names. Simply correcting this appears to have fixed this issue.

All unit and integration tests pass with these changes.