MODE-1146 REST Servlet Ignores Servlet Mappings Other Than /*
Attached patch that corrects the way that the REST server produces URIs when querying the server-level resource and the repository-level resource.  Higher-level resources are not affected by this issue, because we use relative URIs for node children.  The patch remaps the servlet in the tests from /* to /test/* and slightly modifies some test scripts accordingly.