MODE-1122 in case JAAS fails go for anonymous
Attached patch that falls back to using an anonymous user context if one is configured and the attempt to login with simple credentials fails.  If this patch looks good, I can modify the documentation to reflect the new behavior.  This updated version controls this behavior with a Repository Option (on by default).