MODE-1145 Added support for residual properties in queries
It is now possible to use residual properties in JCR-SQL2 queries. For example, if a node is defined with a primary type of "nt:unstructured" and no mixin types, then any properties defined on that node are referred to as 'residual' properties (because they map to 'residual property definitions' on "nt:unstructured", which have names of "*" and/or 'undefined' types).

Prior to this change, residual properties could not be used within a query. Now, however, they can be used just like any other non-residual property. Only those nodes with the property will show up in the results. Note, however, that there is no error if a query uses a property that does not exist.

Several new unit tests were added with these changes. All unit and integration tests pass.