MODE-1147 Added support for vendor extensions in the CND importer
The CND tokenizer already supported capturing vendor extensions as a separate kind of token, but it was not used in the CndImporter. We should enable this functionality in the importer.

However, we don't need to use it in the 'modeshape-jcr' module, as we have no extensions that will be required. First of all, the 'nt:nodeTypeDefinition', 'nt:propertyDefinition', and 'nt:childNodeDefinition' node type definitions do not allow extra properties (we'd have to use a mixin for this). Secondly, we wouldn't use them at all anyway.

All unit and integration tests pass with these changes.