MODE-1089 Test getting version history for newly versionable node
Attempted to replicate the problem reported in MODE-1089 by adding a test case that does exactly the same logic as in the issue. (Thanks Marek for an excellent bug report!)

However, the new test succeeds. I suspect that this issue was fixed with recent changes to rewrite and improve performance of the system graph for versioning. See MODE-1134 for details.