MODE-1050 NPE in NodeTypeManager when validating a recursive node definition
RTNM.registerNodeTypes iterates over the list of node types to be registered, adding each to a list of pending node types after it has been validated.  However, validating a node type that uses itself as a required primary type for a child node definition fails.  This is because validating the required primary types for child nodes checks the list of currently registered node types and the list of pending node types to ensure that the required primary type is in one list or the other, but the node type being validated hasn't been added to the list of pending node types yet because it hasn't been validated.

The fix is to add the node type being validated to the list of pending node types _before_ it gets validated.  If the validation fails, the method throws an exception anyway, so it doesn't matter than an invalid node type definition got into the local list of pending node types.

All tests pass, including the new one that Edouard was kind enough to donate in the MODE-1050 description.