MODE-1156 Client JAR is now included in distribution
The client JAR (which is actually the 'modeshape-jdbc-<version>-jar-with-dependencies.jar' file) is now copied into the 'modeshape-distribution/target' directory and renamed to 'modeshape-<version>-client.jar', and is also placed into the 'modeshape-<version>-dist.zip' file. The release script was also changed to copy the information into the archive directory.