MODE-32 Implemented sequencer for XML Schema Documents
Created a new 'modeshape-sequencer-xsd' module that parses and sequences XSDs. It uses the Eclipse XSD library (which is based on EMF). The XSD sequencer outputs a structure that was designed to mirror the proposed S-RAMP spec (which is currently still in OASIS Technical Committee), and so there is a S-RAMP namespace and a few base node types defined in a new 'modeshape-sequencer-sramp' module.

I had initially tried and would have preferred to use the XSOM parser library (which has a far better API and actually follows Maven best practices), but had severe trouble resolving imports/includes where Eclipse XSD appears to work much better so far.

All unit and integration tests pass. Also, the reference guide was updated with a new chapter for the XSD sequencer.