MODE-1162 Cleaned up unused classes and recent changes
Removed the org.modeshape.web.jcr.rest.model package and the four JAXB binding classes, which are no longer used. Also fixed a catch clause, and moved the 'getRepositoryMetadata(...)' from the AbstractHandler to the ServerHandler class. These were all suggested in the review of the first pull request for this issue.

All unit and integration tests pass.