MODE-1101 The SVN connector does not work properly with "svn://" URL format
THe SVN connector was using different offsets in the directory tree for reads and writes.  As a result, the connector would calculate the wrong path for almost any file or folder that was to be written and default to dropping the file or folder in the root of the repository.  The attached patch corrects this misalignment.

A new test is added, but the test requires the presence of a writable SVN repository - a luxury that is not generally available.