MODE-1015 ModeShape unit tests are failing with DB2 9.7
As noted in the accompanying JIRA, the root cause of this issue is that DB2 (and Derby) use a binary data type for JPA LOBs with a maximum size of 255 unless a different size is explicitly provided.  Most, if not all, non-trivial uses of ModeShape require a larger value for the two LOB fields in the JPA Simple model.

This patch addresses the problem in several ways.  First, it provides larger default values for the two LOB columns (MODE_SIMPLE_NODE.DATA and MODE_LARGE_VALUES.DATA) that ModeShape uses.  Second, it adds comments in the model classes for the JPA entities that contain these fields.  Third, it adds a detailed note in the JPA Connector of the Reference Guide explaining how to address these issues in a production environment.  Fourth, it modifies the DDL generator utility to output a warning when DDL is generated for the DB2 or Derby dialects.