MODE-1179 Add Connector for Disk-Based Storage
Added a patch that implements a connector that supports on-disk storage of arbitrary node content.  The connector currently creates different folder structures for each workspace and creates a file for each node.  Efficient access by UUID is supported.  This patch includes documentation in the Reference Guide and passes all TCK tests.