MODE-947 ModeShape documentation for sequencers needs update
Attached updates to documentation for the sequencers that did not provide some indication of their output format.  I hand-generated the output for the MP3, MS Office, ZIP, and CND sequencers, although I am pretty confident of my output.

The new sequencer output all note the effects of the mode:derived mixin type and its properties.  I'm undecided on whether it makes more sense to pull that out of the output and add a global note at the beginning of the sequencer chapter or retrofit it into the documentation for the other sequencers.  I'm leaning towards the former for efficiency, but would welcome any input.