MODE-1097 ModeShape JPA repository in JBoss AS takes a long time to initialize when number of nodes are large
I added a test case to reproduce the issue that Souvik noted.  The gist of the issue seems to be that the reindexing that happens on engine startup requires a load of the entirety of the content graph.  This exacerbates any existing performance issues with connectors.  Here's the output from the test case for the FileSystemSource:

String built
Reloaded content for / in 00:00:00.001,981
Flushed existing content in 00:00:00.081,379
Reindexed content in 00:00:00.176,902
Initial Startup (creating schema): 00:00:03.065,375
Inserted 110 nodes: 00:00:16.462,787
Restarting engine...
Reloaded content for / in 00:00:00.000,456
Flushed existing content in 00:00:00.000,338
Reindexed content in 00:00:00.004,089
Subsequent Startup (no schema validation): 00:00:00.512,491

And the output for the same test with the JpaSource using an in-memory Derby database:

Reloaded content for / in 00:00:00.256,798
Flushed existing content in 00:00:00.152,776
Reloaded content for / in 00:00:00.057,895
Flushed existing content in 00:00:00.000,603
Reindexed content in 00:00:00.616,951
Initial Startup (creating schema): 00:00:07.848,034
Inserted nodes: 00:00:24.628,916
Restarting engine...
Reloaded content for / in 00:01:00.589,765
Flushed existing content in 00:00:00.000,395
Reloaded content for / in 00:00:00.083,478
Flushed existing content in 00:00:00.000,323
Reindexed content in 00:02:26.031,058
Subsequent Startup (no schema validation): 00:02:26.754,994

Even for an evenly-balanced 3x10 tree, there's a pretty profound difference in startup times.  One short-term workaround would be to place large content in a file-based source when possible.  Another low-impact possibility is to add a repository option that allows administrators to disable the reindexing entirely.  This would be useful for repositories where content is not exposed externally (e.g., JpaStore, InfinispanStore) and the repository isn't clustered.