MODE-33 Added WSDL 1.1 sequencer
Added a new sequencer module for parsing and extracting information from Web Service Definition Language (WSDL) 1.1 files. The structure of the content derived from the WSDL files is based upon the S-RAMP WSDL and SOAPWSDL models, although the structure adds a fair amount of additional information and detail.

Both unit and Integration tests were added for this new sequencer, and the sequencer was added to the JBoss AS kit. All unit and integration tests pass with these changes.

The Reference Guide was updated to describe the new WSDL sequencer, and a few corrects were made to include references to the XSD, WSDL and S-RAMP sequencer modules earlier in the document.