MODE-1172 Corrected OPV behavior of 'checkin'
The JcrVersionManager.checkin(...) method was incorrectly determining whether/how to version child nodes based upon the OPV of the node being checked in, whereas it should have use the OPV of each child node. Exceptions to this rule are spelled out in Section 3.13.9 items 5 and 6, which specify that COPY semantics are to be used for the entire subgraph when a child node has an OPV of COPY or (if the child node is not versionable) VERSION.

The code was changed to reflect the specified behavior. These changes were actually quite minimal. However, several ModeShapeTckTest methods needed to be changed, as they were enforcing the old incorrect behavior.

Note, however, that all of the JCR TCK tests passed with and without these changes. Obviously, they were not thorough enough when it came to these situations.