MODE-1176 Changed ModeShapeUnitTest to correctly find config files on classpath
When the 'startEngineUsing' method was called by a static method, it could not determine the classloader and thus could not find the configuration file that was on the classpath. The solution was to add another form of the method that took the class under test as a second parameter, and to make the 'startEngineUsing(String)' method non-static so that can only be called when the class under test can be determined.