MODE-1180 Starting the JcrEngine starts repositories in parallel
When starting the JcrEngine and requesting that the repositories be initialized now performs the initialization in parallel (up to 4 at a time). There is also now an alternative method that accepts a timeout specification, after which the method will return even if the repositories are not all initialized. Note, however, that in such cases, getRepository(String) will block until all repositories are initialized.

A new unit test was added, and all unit and integration tests pass.