MODE-1180 Improved the locking mechanism of JcrEngine
The locking within JcrEngine to guarantee concurrent access to the repositories was a bit primitive and simplistic - the initialization of all repositories locked out all access to other repositories.
This new approach allows the initialization to proceed without maintaining the locks. In fact, any operation that uses a JcrRepository will lock the engine only when obtaining the JcrRepositoryHolder. Once the holder is obtained, the getting of the JcrRepository will not require the lock, but may block until the repository is initialized.